+++ src/decorators/inject-store.decorator.ts
@@ -90,7 +90,7 @@ export function InjectStore(newPath: string[] | string | ((currentPath, stateInd
     };
 }
 
-export class HasStore<T> {
-    store: Store<T>;
-    state?: any;
+export abstract class HasStore<T> {
+    store: Store<T> = null;
+    state?: any = null;
 }